OneDrive saver SDK for Android
Your app can provide a way for users to save files to OneDrive and
OneDrive for Business. The saver launches a OneDrive experience, the saver
experience, in which the user can choose a location to save the file
provided by your app.

Upload a file of your choice to the folder of your choice.